use warehouse to create lake table and insert
