+++ fe/fe-core/src/main/java/com/starrocks/common/proc/BackendsProcDir.java
@@ -194,12 +194,6 @@ public class BackendsProcDir implements ProcDirInterface {
             backendInfo.add(String.format("%.2f", memUsedPct * 100) + " %");
             backendInfo.add(String.format("%.1f", backend.getCpuUsedPermille() / 10.0) + " %");
 
-            if (Config.integrate_starmgr) {
-                backendInfo.add(String.valueOf(backend.getStarletPort()));
-                long workerId = GlobalStateMgr.getCurrentState().getStarOSAgent().getWorkerIdByBackendId(backendId);
-                backendInfo.add(String.valueOf(workerId));
-            }
-
             comparableBackendInfos.add(backendInfo);
         }
