+++ fe/fe-core/src/main/java/com/starrocks/qe/Coordinator.java
@@ -43,6 +43,7 @@ import com.google.common.collect.Sets;
 import com.starrocks.analysis.DescriptorTable;
 import com.starrocks.authentication.AuthenticationManager;
 import com.starrocks.catalog.FsBroker;
+import com.starrocks.common.Config;
 import com.starrocks.common.MarkedCountDownLatch;
 import com.starrocks.common.Pair;
 import com.starrocks.common.Status;
@@ -72,7 +73,6 @@ import com.starrocks.sql.ast.UserIdentity;
 import com.starrocks.sql.common.ErrorType;
 import com.starrocks.sql.common.StarRocksPlannerException;
 import com.starrocks.system.ComputeNode;
-import com.starrocks.system.SystemInfoService;
 import com.starrocks.task.LoadEtlTask;
 import com.starrocks.thrift.TCompressionType;
 import com.starrocks.thrift.TDescriptorTable;
@@ -489,11 +489,11 @@ public class Coordinator {
             receiver = new ResultReceiver(
                     topParams.instanceExecParams.get(0).instanceId,
                     coordinatorPreprocessor.getAddressToBackendID().get(execBeAddr),
-                    SystemInfoService.toBrpcHost(execBeAddr),
+                    GlobalStateMgr.getCurrentSystemInfo().toBrpcHost(execBeAddr),
                     queryOptions.query_timeout * 1000);
 
             // Select top fragment as global runtime filter merge address
-            setGlobalRuntimeFilterParams(topParams, SystemInfoService.toBrpcHost(execBeAddr));
+            setGlobalRuntimeFilterParams(topParams, GlobalStateMgr.getCurrentSystemInfo().toBrpcHost(execBeAddr));
 
             if (LOG.isDebugEnabled()) {
                 LOG.debug("dispatch query job: {} to {}", DebugUtil.printId(queryId),
@@ -1075,7 +1075,8 @@ public class Coordinator {
                 for (final CoordinatorPreprocessor.FInstanceExecParam instance : params.instanceExecParams) {
                     TRuntimeFilterProberParams probeParam = new TRuntimeFilterProberParams();
                     probeParam.setFragment_instance_id(instance.instanceId);
-                    probeParam.setFragment_instance_address(SystemInfoService.toBrpcHost(instance.host));
+                    probeParam.setFragment_instance_address(
+                            GlobalStateMgr.getCurrentSystemInfo().toBrpcHost(instance.host));
                     probeParamList.add(probeParam);
                 }
                 if (coordinatorPreprocessor.isUsePipeline() && kv.getValue().isBroadcastJoin() &&
@@ -1828,7 +1829,7 @@ public class Coordinator {
                     return false;
                 }
 
-                TNetworkAddress brpcAddress = SystemInfoService.toBrpcHost(address);
+                TNetworkAddress brpcAddress = GlobalStateMgr.getCurrentSystemInfo().toBrpcHost(address);
 
                 try {
                     BackendServiceClient.getInstance().cancelPlanFragmentAsync(brpcAddress,
@@ -1857,6 +1858,11 @@ public class Coordinator {
         }
 
         public boolean isBackendStateHealthy() {
+            if (Config.use_staros) {
+                // get worker state from SystemInfo Service
+                return GlobalStateMgr.getCurrentSystemInfo().checkWorkerHealthy(backend.getId());
+            }
+
             if (backend.getLastMissingHeartbeatTime() > lastMissingHeartbeatTime) {
                 LOG.warn("backend {} is down while joining the coordinator. job id: {}", backend.getId(), jobId);
                 return false;
