+++ fe/fe-core/src/main/java/com/starrocks/system/HeartbeatMgr.java
@@ -142,9 +142,11 @@ public class HeartbeatMgr extends LeaderDaemon {
         }
 
         // send compute node heartbeat
-        for (ComputeNode computeNode : nodeMgr.getIdComputeNode().values()) {
-            BackendHeartbeatHandler handler = new BackendHeartbeatHandler(computeNode);
-            hbResponses.add(executor.submit(handler));
+        if (!Config.use_staros) {
+            for (ComputeNode computeNode : nodeMgr.getIdComputeNode().values()) {
+                BackendHeartbeatHandler handler = new BackendHeartbeatHandler(computeNode);
+                hbResponses.add(executor.submit(handler));
+            }
         }
 
         // send frontend heartbeat
@@ -241,15 +243,6 @@ public class HeartbeatMgr extends LeaderDaemon {
                         if (!isReplay && !computeNode.isAlive()) {
                             GlobalStateMgr.getCurrentState().getGlobalTransactionMgr()
                                     .abortTxnWhenCoordinateBeDown(computeNode.getHost(), 100);
-                        }
-                    } else {
-                        if (Config.integrate_starmgr && !isReplay) {
-                            // addWorker
-                            int starletPort = computeNode.getStarletPort();
-                            if (starletPort != 0) {
-                                String workerAddr = computeNode.getHost() + ":" + starletPort;
-                                GlobalStateMgr.getCurrentState().getStarOSAgent().addWorker(computeNode.getId(), workerAddr);
-                            }
                         }
                     }
                     return isChanged;
@@ -452,6 +445,9 @@ public class HeartbeatMgr extends LeaderDaemon {
     }
 
     public void replayHearbeat(HbPackage hbPackage) {
+        if (Config.use_staros) {
+            return;
+        }
         for (HeartbeatResponse hbResult : hbPackage.getHbResults()) {
             handleHbResponse(hbResult, true);
         }
