+++ fe/fe-core/src/test/java/com/starrocks/sql/plan/PlanFragmentWithCostTest.java
@@ -1548,7 +1548,7 @@ public class PlanFragmentWithCostTest extends PlanTestBase {
             throws Exception {
         new Expectations() {
             {
-                GlobalStateMgr.getCurrentSystemInfo().isSingleBackendAndComputeNode();
+                GlobalStateMgr.getCurrentSystemInfo().isSingleBackendAndComputeNode(0);
                 returns(true, true, true, false, true);
             }
