+++ fe/fe-core/src/test/java/com/starrocks/system/SystemInfoServiceTest.java
@@ -50,7 +50,7 @@ public class SystemInfoServiceTest {
             SecurityException,
             IllegalArgumentException,
             IllegalAccessException {
-        service = new SystemInfoService();
+        service = new LocalSystemInfoService();
         Field field = FrontendOptions.class.getDeclaredField("useFqdn");
         field.setAccessible(true);
         field.set(null, true);
